Build POC for rendering a content page
The POC is a combination of the ember-mobiledoc-dom-renderer and the
ember-mobiledoc-editor addons.  The combination of these two allows us
to edit and display the contents of the website.  All in all, this
should be sufficient for us to build an editable site.

Necessary input, which hasn't been reviewed sufficiently, has been taken
from [1].  It should be reviewed and possibly abstracted.  We suspect
support for this style of component to arrive in the
ember-mobiledoc-dom-renderer library.

[1] https://github.com/bustlelabs/mobiledoc-kit/blob/5b887b1/demo/app/components/mobiledoc-dom-renderer.js