Introducing editable-page and using it
The editable-page is a simple component that allows to register a page
which aught to be editable.  It either renders the editor, or the
rendered content, based on a toggle button.  Should be a good enough
start for our first editable page support.

We have removed the unnecessary contents from the
applicationcontroller in order to display this stuff.  We may revive it
as we work out the details of the editable page.