Create route for displaying pages and editing pages
The edit pages are separated from the display pages.  This would allow
us to simplify the displaying later on.  For now, they still use very
similar logic.

Content can be saved in mu-cl-resources as expected and pages can be
edited as necessary.  Still needed is a menu-structure, more components
for editing and some edits in the current components.